TAMAYA-327: Consistent signature creating ConversionContexts
Comments in ConversionContext lead me to believe that supportedFormats
should have their order maintained after import.  The HashSet that the
Builder object uses as a backer will not accomplish that goal.  This
change converts the Builder to use a LinkedHashSet supportedFormats, as
well as the object proper.

Some additional tests have been added here in pursuit of TAMAYA-288's
additional code coverage.