TAMAYA-328: Let CharConverter parse a single apostrophe character
CharConverter uses paired apostrophes (') to delimit multi-character
strings and pulls the first into a Character object as needed.  That
works great, but has made it to where a single apostrophe by itself
cannot be used directly.  Entertainingly, "'''" works, but "'" throws an
exception.  This adds a simple check if the string is 1 character long
after having determined that it starts with an apostrophe.