TAMAYA-340 Bugfix for regex "\\." vs literal "."
String.replaceAll(String, String) takes a regex as its first argument,
not a literal replaceable string.  The result of this is that any
single-character environment variable is converted to use the "_" key
which has special significance to the system.  Tests in the extensions
repo include some that have a single character which this change should
get working again.  I've included a test in this repo for that case as
well.