TAMAYA-383 Let OSGi skip thread classloader key
We use the thread's classloader as a key into a hashmap to find the
serviceloader for a given configuration.  In OSGi, we have separate
classloaders for the CoreConfigurationBuilder,
DefaultConfigurationBuilder, and ServiceContextManager.  That confusion
prevents the simple case from finding a match in the hashmap that can
load a particular ServiceContext.  Since
CoreConfiguation/DefaultConfiguration already uses the static class's
classloader during <init>, this PR allows the OSGIActivator to point
that classloader at the relevant OSGIServiceContext.